admin: change angular to be role based. Dashboard: change user status. profile: soft delete account. otp: add timeout and resend otp. landing page: created.

+++ financial-dashboard-node/src/repositories/otp.repository.ts
@@ -7,13 +7,22 @@ const otpRepository = AppDataSource.getRepository(OTP);
 export const generateOtp = async(data:Partial<OTP>)=>{
     try{
         const otp = Math.floor(100000+Math.random()*900000).toString();
+        console.log(otp);
+        
         let date = new Date();
-        let result = await otpRepository.save({
-            email:data.email,
-            otp:otp,
-            createdAt:date,
-            type:data.type,
-        })
+        let existing = await otpRepository.findOne({where:{email:data.email}})
+
+        if(existing){
+            await otpRepository.update({email:data.email},{otp:otp,createdAt:data.createdAt});
+        }
+        else{
+            let result = await otpRepository.save({
+                email:data.email,
+                otp:otp,
+                createdAt:date,
+                type:data.type,
+            })
+        }
 
         return otp;
     }
@@ -25,10 +34,17 @@ export const generateOtp = async(data:Partial<OTP>)=>{
 export const verifyOtp = async(data:Partial<OTP>)=>{
     try{
         let result = await otpRepository.findOneBy({email:data.email,otp:data.otp,type:data.type});
+        
         if(result===null){
             throw new AppError('Wrong OTP',500);
         }
         else{
+            let timeLapsed = Date.now()-result.createdAt.getTime();
+            console.log(timeLapsed);
+            
+            if(timeLapsed/1000 >180){
+                throw new AppError('OTP invalid',500);
+            }
             await otpRepository.delete({email:data.email})
             return result;
         }
